+++ CHANGELOG.md
@@ -18,6 +18,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - `NodeDistribution`: Concept for constraining the template parameter of consumers of node distributions.
 - `DensityDistribution`: Class template for the distribution of density over a velocity set.
 
+- `EquilibriumDistributionBase`: Base class for equilibrium distributions.
+- `EquilibriumDistribution`: Concept for constraining the template parameter of consumers of equilibrium distributions.
+- `DiscreteMaxwellBoltzmann`: Class template for the discrete Maxwell Boltzmann distribution approximated via a Hermite expansion. Currently only the second-order approximation is implemented.
+
 - GitHub workflow for continuous integration (CI)
 - Ubuntu 24.04 Dockerfile for installing dependencies needed CI workflow
 - CMake target `Coverage` for running code coverage analysis